+++ src/lib.rs
@@ -34,6 +34,7 @@ mod config_load_balance;
 mod config_loader;
 mod config_logging;
 mod config_net;
+mod config_observability;
 mod config_path;
 mod config_php;
 mod config_route;